[REF][SLOT-283]: add refactor due to code review
