+++ src/main/java/com/three_tech_solutions/slot_app/components/monthly_fee_processors/implementations/BeginningOfMonthMonthlyFeeProcessor.java
@@ -10,7 +10,7 @@ import org.springframework.web.server.ResponseStatusException;
 import java.time.LocalDate;
 import java.time.Month;
 
-import static com.three_tech_solutions.slot_app.utils.PaymentPlanUtils.BEGINNING_OF_MONTH_EXPIRATION_DATE;
+import static com.three_tech_solutions.slot_app.constants.PaymentPlanConstants.BEGINNING_OF_MONTH_EXPIRATION_DATE;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 
 @Component
